Apply PR-derived hardening bundle across daemon, db, runtime, alpaca, GUI, and DB proof tooling

+++ core-rs/crates/mqk-execution/src/gateway.rs
@@ -343,7 +343,7 @@ where
         &self,
         internal_id: &str,
         order_map: &BrokerOrderMap,
-        quantity: i32,
+        quantity: i64,
         limit_price: Option<i64>,
         time_in_force: String,
     ) -> Result<BrokerReplaceResponse, Box<dyn std::error::Error>> {
